backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pandas
numpy
scikit-learn
xgboost
joblib
